put the function poketype in poketypefunction
